Add useConnectedInputs hook for workflow inputs
