+++ internal-packages/workflow-designer-ui/src/editor/properties-panel/action-node-properties-panel/lib/index.ts
@@ -1,1 +1,2 @@
 export * from "./connected-outputs";
+export * from "./use-connected-inputs";
